Check against response status codes
Set the .good / .bad classes to .good only on 2XX response statuses.

Don't only check for a response; check the status code as well. A status
code outside of the 2XX range indicates either an unexpected setup (in
the case of 3XX or 4XX responses) or a server outage (in the case of 5XX
responses).